+++ packages/apple_intelligence/lib/apple_intelligence.dart
@@ -6,3 +6,5 @@ export 'src/concept.dart';
 export 'src/exceptions.dart';
 export 'src/image_creator.dart' show GeneratedImage, ImageCreator;
 export 'src/image_playground_sheet.dart' show ImagePlaygroundSheet;
+export 'src/native_text_field.dart'
+    show AppleIntelligenceTextField, NativeTextController, TextCapabilities;